backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
feedparser==6.0.11
beautifulsoup4==4.12.3
rapidfuzz==3.9.7
reportlab==4.2.2
aiosmtplib==3.0.2
aioimaplib==1.1.0
celery[redis]==5.4.0
redis==5.1.1
boto3==1.35.19
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
anthropic==0.34.2
structlog==24.4.0
uuid7==0.1.0
Pillow==10.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
